# docs/swagger.yaml
basePath: /api
definitions:
...
  handler.AcceptInterestBody:
    properties:
      phone:
        type: string
    type: object
  handler.AcceptInterestResponse:
    properties:
      searcher_phone:
        type: string
    type: object
...
  handler.ContactInfo:
    properties:
      departure_at:
        type: string
      destination:
        type: string
      name:
        type: string
      origin:
        type: string
      phone:
        type: string
      role:
        type: string
    type: object
...
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handler.ExpressInterestBody:
    properties:
      name:
        type: string
      phone:
        type: string
    type: object
  handler.ExpressInterestResponse:
    properties:
      id:
        type: string
      status:
        type: string
    type: object
  handler.FeedbackBody:
    properties:
      phone:
        type: string
      taken:
        type: boolean
    type: object
...
  handler.MyInterest:
    properties:
      departure_at:
        type: string
      destination:
        type: string
      driver_name:
        type: string
      id:
        type: string
      origin:
        type: string
      ride_id:
        type: string
      status:
        type: string
    type: object
  handler.NotificationItem:
    properties:
      departure_at:
        type: string
      destination:
        type: string
      driver_name:
        type: string
      origin:
        type: string
      request_id:
        type: string
      ride_id:
        type: string
      sent_count:
        type: integer
    type: object
...
info:
  contact: {}
  description: Local ride-sharing notice board API.
  title: Go-Stop API
  version: "1.0"
paths:
...
  /destinations:
    get:
      operationId: listDestinations
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      tags:
      - destinations
  /interests:
    get:
      operationId: listMyInterests
      parameters:
      - description: Searcher phone
        in: header
        name: X-Phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.MyInterest'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      tags:
      - interests
  /interests/{id}/accept:
    post:
      consumes:
      - application/json
      operationId: acceptInterest
      parameters:
      - description: Interest ID
        in: path
        name: id
        required: true
        type: string
      - description: Driver phone
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.AcceptInterestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.AcceptInterestResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      tags:
      - interests
  /interests/{id}/contact:
    get:
      operationId: getInterestContact
      parameters:
      - description: Interest ID
        in: path
        name: id
        required: true
        type: string
      - description: Caller phone
        in: header
        name: X-Phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ContactInfo'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      tags:
      - interests
  /notifications:
    get:
      operationId: listNotifications
      parameters:
      - description: Searcher phone
        in: header
        name: X-Phone
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.NotificationItem'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      tags:
      - notifications
...
  /rides/{id}/feedback:
    post:
      consumes:
      - application/json
      operationId: submitRideFeedback
      parameters:
      - description: Ride ID
        in: path
        name: id
        required: true
        type: string
      - description: Owner phone and taken flag
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.FeedbackBody'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      tags:
      - rides
  /rides/{id}/interest:
    post:
      consumes:
      - application/json
      operationId: expressInterest
      parameters:
      - description: Ride ID
        in: path
        name: id
        required: true
        type: string
      - description: Searcher phone and name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.ExpressInterestBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.ExpressInterestResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      tags:
      - interests
...
swagger: "2.0"